added back button implementation, player fragment renamed
