+++ app/src/main/res/layout/fragment_player_small.xml
@@ -4,7 +4,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:orientation="vertical"
-                tools:context=".PlayerFragment">
+                tools:context=".SmallPlayerFragment">
 
     <ProgressBar
             android:id="@+id/fragment_progressbar"
